+++ CHANGELOG.md
@@ -84,6 +84,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix bug for AI + tool calls.
 - [frontend] - Fix docker compose redirect bug.
 - [frontend] - Refactor
+- Websearch tools.
 
 ## [0.3.3] - 30.10.2024